+++ include/ftxui/dom/elements.hpp
@@ -81,6 +81,7 @@ Element color(Color, Element);
 Element bgcolor(Color, Element);
 Decorator focusPosition(int x, int y);
 Decorator focusPositionRelative(float x, float y);
+Element automerge(Element);
 
 // --- Layout is
 // Horizontal, Vertical or stacked set of elements.