+++ src/ftxui/component/dropdown.cpp
@@ -17,8 +17,8 @@ Component Dropdown(ConstStringListRef entries, int* selected) {
     Impl(ConstStringListRef entries, int* selected)
         : entries_(std::move(entries)), selected_(selected) {
       CheckboxOption option;
-      option.style_checked = "↓│";
-      option.style_unchecked = "→│";
+      option.style_checked = "↓ ";
+      option.style_unchecked = "→ ";
       checkbox_ = Checkbox(&title_, &show_, option),
       radiobox_ = Radiobox(entries_, selected_);